Changing files, should have updated the orignal file, not had a new refrecnce/test file, See below same commit message explaining tests
This is the real test commit, Copying most of the origianl testing but changing some details.
First this is that I took the orginal test as the first two tests were basically
done. I improved them by adding in a second part of the test to make sure that
the function returns a float as that is an aspect of the function. Then I added
two more tests, One being a postitive number to make sure it was still a postivie number.
Next was not specifically in the spec, but it more of an intutive error, we should
get a type error when the function is given a non-number or number (float, Long, int, ect.).
